third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/exceptions-expected.txt

Test that XMLHttpRequest raises exceptions when it should.

new XMLHttpRequest()
PASS: req.setRequestHeader("Foo", "bar") threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS: req.send(null) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
open()
PASS: req.setRequestHeader() threw exception TypeError: Not enough arguments.
PASS: req.setRequestHeader("Foo") threw exception TypeError: Not enough arguments.
PASS: req.status() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS: req.statusText() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
send()
PASS: req.send(null) threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS: req.setRequestHeader("Foo", "bar") threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS: req.getResponseHeader() threw exception TypeError: Not enough arguments.
PASS: req.open() threw exception TypeError: Not enough arguments.
PASS: req.open(null) threw exception TypeError: Not enough arguments.
